requirements.txt
# Database
# SQLite is built-in Python 3.x, no driver needed
sqlalchemy==2.0.25

# AI Framework (newer versions, OpenAI SDK v1)
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
openai==1.51.0

# HTTP client
# NOTE: openai==1.51.0 passes `proxies=` to httpx.Client, which was removed in httpx>=0.28.
# Pinning to 0.27.x avoids the `Client.__init__() got an unexpected keyword argument 'proxies'` error.
httpx==0.27.2

# Frontend
streamlit==1.30.0

# Data Generation & Analysis
faker==22.0.0
pandas==2.1.4

# Utils
python-dotenv==1.0.0
pydantic==2.5.3

# Optional: Only if migrating to PostgreSQL
# psycopg2-binary==2.9.9
